+++ src/main/java/net/silentchaos512/torchbandolier/TorchBandolier.java
@@ -21,8 +21,8 @@ import net.silentchaos512.torchbandolier.init.ModItems;
 public class TorchBandolier implements IModBase {
     public static final String MOD_ID = "torchbandolier";
     public static final String MOD_NAME = "Torch Bandolier";
-    public static final String VERSION = "0.1.1";
-    public static final String VERSION_SILENTLIB = "3.0.0";
+    public static final String VERSION = "1.0.0";
+    public static final String VERSION_SILENTLIB = "3.0.6";
     public static final int BUILD_NUM = 0;
     public static final String DEPENDENCIES = "required-after:silentlib@[" + VERSION_SILENTLIB + ",)";