+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-TODO
+### Added
+- Config for max torch count
 
 ## [0.1.1] - 2018-08-19
 Updated for Silent Lib 3.0.0